% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_test.R
\name{area_estimation}
\alias{area_estimation}
\title{Area estimation for projected rasters}
\usage{
area_estimation(pred, cnf, conf = 1.95)
}
\arguments{
\item{pred}{A categorical raster}

\item{cnf}{A confusion matrix from \code{link[caret]{confusionMatrix}}.}

\item{conf}{A confidence interval}
}
\value{
A table with columns cell with original and adujsted area estiamates
per class as well Producer's, Users's and Overall accuracies with confidence
interval.
}
\description{
Area is estimated based on the pixel resolution. Projected CRS should be
used of meaningful area estimation because non-projected CRS, e.g. geographic
coordinates do not have regular cell size.
}
\references{
Olofson P., Foody G.M., Herold M., Stehman S.V., Woodcock C.E., Wulder M.A. 2014.
Good practices for estimating area and assessing accuracy of land change
\url{https://doi.org/10.1016/j.rse.2014.02.015}
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
