% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_test.R
\name{eval_test}
\alias{eval_test}
\title{Evaluate a testing set}
\usage{
eval_test(model, testset, response)
}
\arguments{
\item{model}{A model object.}

\item{testset}{A data.frame containing the test data to predict}

\item{response}{A character vector indicating the column name of the response
variable.}
}
\value{
A list with a confusion matrix and the test set amended by a column called pred
containing the predicted values.
}
\description{
Based on a model returned by \code{link{mapme.forest}[train_model]} and a
test set a confusion matrix is returned.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
