% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_traindata}
\alias{extract_traindata}
\title{Extract pixel values from a predictor set per polygon}
\usage{
extract_traindata(aoi, predictors, verbose = TRUE, ...)
}
\arguments{
\item{aoi}{A sf object with polygons for which pixel values are extracted.}

\item{predictors}{A SpatRaster object for each layer the pixel values are
extracted. The column names of the resulting object will be based on the
layer names.}

\item{verbose}{A logical indicating the level of verbosity.}

\item{...}{Additional parameters used by \code{\link[terra]{extract}}.}
}
\value{
matrix, list, or data.frame
}
\description{
This function is used to extract pixel values of a predictor set of type
SpatRaster on a polygon basis of a sf object. The CRS of the two objects
should be the same, however, the function will try to transform the aoi
object if that is not the case. See \code{\link[terra]{extract}} for fine
controll of the extraction process.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
