% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_test.R
\name{postclass}
\alias{postclass}
\title{Post-classification by focal operation}
\usage{
postclass(pred, w, fun = "modal", ...)
}
\arguments{
\item{pred}{A SpatRaster object corresponding to the prediction}

\item{w}{The window can be defined as one (for a square) or two numbers (row, col);
or with an odd-sized weights matrix. See \code{link{terra}[focal]}.}

\item{fun}{function that takes multiple numbers, and returns a single number.
For example mean, modal, min or max. It should also accept a na.rm argument,
either as actual argument or through use of ... See \code{link{terra}[focal]}.}

\item{...}{Additional parameters used by See \code{link{terra}[focal]}.}
}
\value{
A SpatRaster object.
}
\description{
Uses terra::focal to apply a post-classification based on a spatial moving
window.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
