% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_test.R
\name{spatial_predict}
\alias{spatial_predict}
\title{Spatial prediction of a trained model}
\usage{
spatial_predict(
  model,
  predictors,
  AOA = TRUE,
  mode = "raw",
  ncores = 1,
  mask = NULL,
  ...
)
}
\arguments{
\item{model}{A model object.}

\item{predictors}{A spatRaster with the predictor variables.}

\item{AOA}{A logical indicating if the Area of Applicability shall be calculated.
Defaults to FALSE. Can be quite a costly calculation. Check the documentation of
\code{link{CAST}[aoa]} for possible speed ups via the ... arguments, i.e.
through parallel computation.}

\item{mode}{A character vector indicating the behavior of \code{link{terra}[predict]}.
If \code{type = "raw"} one raster layer with the predicted class is returned.
If the model is a regression model the type cannot be set to "proba".
If \code{type = "proba"} for each outcome class a single layer is returned
indicating the probability of a pixel to be classified as that respective class.}

\item{ncores}{An integer specifying the number of cores to be used for parallel computations.}

\item{mask}{A spatRaster object with at the same resoultion as the predictor stack.
All raster cells unqual to NA will be masked out of the predictor stack to
omit predictions at these locations}

\item{...}{Additional arguments used for \code{link{terra}[predict]} and \code{link{CAST}[aoa]}.}
}
\value{
A SpatRaster object with at least one layer which is the classification
result. Depending if and how the AOA is calculated two additional layers are
possible representing the AOA itself and the DI. Check \code{link{CAST}[aoa]}
for more information.
}
\description{
This function is used to spatially predict a trained model on a SpatRaster
object with the predictors. Additionally users can opt in to calculate the
Area of Applicability based on \code{link{CAST}[aoa]}.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
