% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Trains a model possibly based on space-time folds}
\usage{
train_model(
  traindata,
  predictors,
  response,
  spacevar = NULL,
  timevar = NULL,
  k = 10,
  ffs = TRUE,
  method = "rf",
  metric = ifelse(is.factor(response) | is.character(response), "Accuracy", "RMSE"),
  maximize = ifelse(metric == "RMSE", FALSE, TRUE),
  trControl = caret::trainControl(),
  seed = 42,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{traindata}{A tibble, data.frame or sf object.}

\item{predictors}{A character vector with the column names used as predictors.}

\item{response}{A character vector with the name of the response variable.}

\item{spacevar}{An optional character with a variable indicating spatial
groups to account for during training.}

\item{timevar}{An optional character with a variable indicating temporal
groups to account for during training.}

\item{k}{An integer value indicating the number of folds. Defaults to 10.}

\item{ffs}{A logical indicating if a Forward-Feature-Selection should be
conducted. Note that this could significantly increase training time.}

\item{method}{A charachter vector indicating a model to be used by caret.
Defaults to \code{"rf"}.}

\item{metric}{An accuracy metric for model optimization. Defaults to Accuracy
for classification and RMSE for regression. Must be supported by caret.}

\item{maximize}{A logical indicating if the selected metric should be
maximized. Defaults to TRUE for Accuracy and FALSE for RMSE.}

\item{trControl}{A train control object based on \code{\link[caret]{trainControl}}.}

\item{seed}{An integer value used to ensure reproducibility.}

\item{verbose}{A logical indicating the level of verbosity.}

\item{...}{Additional parameters used by \code{\link[CAST]{ffs}} and \code{\link[caret]{train}}.}
}
\value{
A list with the model, performance metrics or confusion matrix (for
regression or classification), a vector of observed values and a vector
of predicted values.
}
\description{
This function is basically a wrapper around functionality from the CAST and
caret packages. It expects as input a data.frame object with the predictors
and outcome variables present and well as possible variables indicating the
spatial and temporal sub-groups of the training data. Either one of the
spatial or temporal variables can be used to create individual space or time
folds
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
