% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_split.R
\name{train_split}
\alias{train_split}
\title{Splitting polygons into training and test}
\usage{
train_split(
  aoi,
  idcol,
  response,
  group = NULL,
  p,
  seed = round(runif(1, 1L, 1000L)),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{aoi}{A sf object with polygons which are to be split into training and
test. The object should contain one column which uniquely identfies the
polygons, a column with a response variable and potentially a additional
grouping variable used for stratification.}

\item{idcol}{A character vector indicating the name of the column that uniquely
identifies the polygons.}

\item{response}{A character vector indicating the name of the column that
contains the response variable which can be either numerical or charachter/factor.
If it is a numerical variable it is mandatory to specify the behavior of
\code{\link[base]{cut}} to discretize the value range into categories.}

\item{group}{A optional character identifying the name of a column used for
stratification of the training-test split. The values of that column need to
be charachters or factors since grouping by numeric variables currently is not
supported.}

\item{p}{A numeric between 0 and 1 indicating the fraction of the polygons which
shall enter the training set.}

\item{seed}{A numeric value used to ensure reproducibility of the split.}

\item{verbose}{A logical indicating the level of verbosity.}

\item{...}{Additional parameters used by \code{\link[base]{cut}} to discretize
the value range of a numeric response variable into categories.}
}
\value{
The original sf object amended by a column named \code{"split"} which
indicates if a polygon belongs to the training or testing data set.
}
\description{
This function is used to split a data set of polygons into two data sets
used for training and testing. Generally the splitting is done based on the
response variable to achieve similar class distributions between the training
and testing set. Additionally, a second grouping variable can be used to
achieve stratification based on that group. The function works on categorical
and numerical data. For numeric variables the value range will be categorized
with \code{\link[base]{cut}} function and users are required to specify the
value breaks.
}
\author{
Darius Görgen (MapTailor Geospatial Consulting GbR) \email{info@maptailor.net}
\cr
\emph{Maintainer:} MAPME-Initiative \email{contact@mapme-initiative.org}
\cr
\emph{Contact Person:} Dr. Johannes Schielein
\cr
\emph{Copyright:} MAPME-Initiative
\cr
\emph{License:} GPL-3
}
